Move DefaultNamespaceOptions to run_common.go
